chore(api): moving routes to package representing api version
Configure both `/api` and `/api/v0` to be handled by the same
proxy/router, and apply the same behavior for `/internal` and
`/internal/v0`. This ensures that all incoming requests to either the
unversioned or v0 path are routed to the identical backend handler,
simplifying maintenance and avoiding duplicate logic.
