+++ api/routes/v0/auth_test.go
@@ -1,4 +1,4 @@
-package routes
+package v0
 
 import (
 	"crypto/rand"
