+++ api/routes/v0/device.go
@@ -1,4 +1,4 @@
-package routes
+package v0
 
 import (
 	"net/http"
