+++ api/routes/v0/healthcheck.go
@@ -1,4 +1,4 @@
-package routes
+package v0
 
 import (
 	"net/http"
