+++ api/routes/v0/session.go
@@ -1,4 +1,4 @@
-package routes
+package v0
 
 import (
 	"net/http"
