+++ api/routes/v0/setup_test.go
@@ -1,4 +1,4 @@
-package routes
+package v0
 
 import (
 	"errors"
