+++ api/routes/v0/stats.go
@@ -1,4 +1,4 @@
-package routes
+package v0
 
 import (
 	"net/http"
