+++ api/routes/v0/user_test.go
@@ -1,4 +1,4 @@
-package routes
+package v0
 
 import (
 	"encoding/json"
